EEG/py.BI.EEG.2014a-GIPSA/requirements.txt
pyriemann==0.2.5
matplotlib==3.0.2
tqdm==4.28.1
pandas==0.20.1
mne==0.16.1
scipy==0.19.0
numpy==1.12.1
scikit_learn==0.21.3
PyYAML==5.1.2
